requirements.txt
# TEMPO NASA Data Visualization Requirements
harmony-py>=1.0.0
xarray>=0.20.0
netCDF4>=1.5.0
matplotlib>=3.5.0
cartopy>=0.20.0
numpy>=1.20.0
streamlit>=1.35.0
plotly>=5.24.0
pandas>=2.0.0
requests>=2.32.0
statsmodels>=0.14.0
python-dotenv>=1.0.0

